svg: the rasterizer moves below the exporter, so non-export consumers can reach it
Rasterization shipped as a pre-export pass, which is the right shape for
rewriting a SceneGraph's texture paths and is unchanged here. What was in the
wrong place was the renderer underneath it.

Every consumer holding a Material and a path sits *beside* the exporter rather
than after it. Studio's live viewport already reached back into mogen-export
for is_svg/render_svg/resolve_svg_size, and an engine that consumes the lowered
SceneGraph and decodes textures itself could not reach them at all without
linking the whole exporter — oxipng, fbxcel and meshopt included — for three
pure functions over `(bytes, size, wrap)`.

So the policy moves to a new `mogen-svg`: mogen-core plus resvg/usvg, both pure
Rust, and it cross-compiles to wasm32-unknown-unknown (checked). The pass stays
in mogen-export, which re-exports the four public items so `mogen_export::is_svg`
and friends keep resolving and Studio needs no edit. `textures-svg` becomes
`dep:mogen-svg`.

There is still exactly one implementation of "what pixels does this SVG
produce", which was the point of the original split; it is simply reachable now.

Verified: -p mogen-svg (12), -p mogen-export (5 svg unit + 27 svg_textures),
-p mogen-core, -p mogen-dsl (663), cargo check --workspace --all-targets, and
cargo check -p mogen-svg --target wasm32-unknown-unknown.
